airflow/requirements.txt
yfinance
apache-airflow-providers-google
pyarrow
dbt-core
dbt-bigquery
